refactor(state): Centralize write barrier & mutex in SafeBoxStateManager
